+++ node/service/src/lib.rs
@@ -7,5 +7,6 @@
 #[cfg(feature = "frequency-no-relay")]
 pub mod block_sealing;
 pub mod chain_spec;
+pub mod common;
 pub mod rpc;
 pub mod service;
