+++ node/service/src/service.rs
@@ -192,7 +192,7 @@ pub async fn start_parachain_node(
 	para_id: ParaId,
 	hwbench: Option<sc_sysinfo::HwBench>,
 ) -> sc_service::error::Result<(TaskManager, Arc<ParachainClient>)> {
-	use crate::block_sealing::convert_address_to_normalized_string;
+	use crate::common::convert_address_to_normalized_string;
 	use common_primitives::offchain::OcwCustomExt;
 	use sc_client_db::Backend;
